Print TLS version and cipher suite.
This change causes the TLS version and cipher suite to be printed when a
TLS handshake is completed. Since these strings don't come from
crypto/tls, there is a possibility that advances in that package will
add versions and cipher suites that aren't recognised here. In that case
"unknown" will be printed and this package will need to be updated.

Fixes xmpp-client #53.